Molding the Database (#9)
* Added a Docs, Data, and Images folders for any future imgs, documentation and data that we may use for this project.

`GlamWireDb.sql` added to the project and has been executed. Alteration may occur at a later date, the NPC foreign key may need to be added to the cases table for referencing.

* Added a new query that contains the alterations requested. After meeting with my project partner we discussed alterations and needs for the program. After brainstorming, NPCEmail is not needed as we will use usernames/fname/lname for looking up contacts, No changes to InboxMessages.

In the Alteration Query the `Upgrades` Table has been updated so that we can insert images into the database using `VARBINARY` as the data type so that we should be able to hold png, jpeg ... etc