Currently two redirect work, the New Game, Load Game and Exit Button. None of the Back buttons work, right now. Still there is quite a bit of placeholder data.
I may keep the label on the new game form hard coded there's no need to alter that really. My next task is to focus on populating the test save data #if debug/ #end if to insert the test data that we can see for now but not in the final program.

I think thats a good place to break for now :)