Kourtnie1 Data (#22)
* restructure change

* A models folder was created to I can speed run making my classes for the program.

The properties have been added to the SaveFile.cs and I will continue on to cases, npc,,,etc.

* NPC model created, This covers the NPC's. There are just properties added...on to the next

* Updated the descriptions for the classes NPC, SaveFile, && Cases. The Classes have been updated

* line deleted.

*Scaffolded* App.config added to use SQLConnection. The connection string has been placed, so that I can retreieve information from the database.

* GlamwireDB class made for the connection string to be added

* referenceing **Pokedex** project-- (see repos),

Using that to work on inserting the DB connection.

* I didn't have some of my NuGet packages installed on this device so I reinstalled them and we are trucking on.

* I needed to retrieve the cases and NPC's from the database. I have placeholder methods to get them that return null for now.

* Altering my try/catches in the `GlamwireDb` so that they will actually work soon.

I need a while loop for it and to copy a generic error message into the catch method.

the null returns will be changed as I go along, I just didn't like having the squiggle

* updated the RetrieveAllNPCs method, bridging the database to retrieve the NPCS and put them in list form based on the active case.

* Going back and forth between making a custom data tye of SaveFile for the Case and NPC classes.

Made my conversions easier for myself to read, a new method to be made and more logic was implemented here.

* I had an invalid data type in the RetrieveActiveCase() method, it is now fixed but I need my new method in to get rid of my remaining error in that method.

* return type changed.

* changed more of my conversion for readability

* Placeholder data for the GetNPCsForCase()  method.

* Documentation added to the `GlamwireDb.cs`,  `SaveFile.cs`, && `Case.cs`

* PlaynameForm added so that the player can start a newgame and enter their "credentials"

* Renamed the Main Glamwire form for simplicity. The previous name seemed crowded

* A new form has been created, there is no continuation data so new game save data will be needed later on. main focus is populating data right now.

Some placeholders have been added to the *new* PlayerNameForm- this is where we will retrieve the playername and transition into a new game

* Currently two redirect work, the New Game, Load Game and Exit Button. None of the Back buttons work, right now. Still there is quite a bit of placeholder data.

I may keep the label on the new game form hard coded there's no need to alter that really. My next task is to focus on populating the test save data #if debug/ #end if to insert the test data that we can see for now but not in the final program.

I think thats a good place to break for now :)

* Allof the current back buttons were given logic to go back to the previous screen. the savefile screen will have to be changed later so that instead of going back to the main screen itjust goes back to whatever screen was before it.

I am going to go ahead and pull request and work form by form && populate data.

* changed the name lets seeif that helps.. rename later.

* Changed the name once more in hopes that the build will pass but I may need to modify GitHubActions

* the button names were all swapped, please keep them cohesive....

pull requesting now

* changing my consolewrite lines to show a message box instead, so that I can havea pop-up ifthe database connection isn't working.

I also deleted my App.config it was just too much work right now to work around it.

* changed my exceptions in the `GlamwireDb` to SQLExceptions!!

* Filled in the blanks to my GetNPCForCase() method, this will retrieve the NPC for the current case the user is working on

* fixed curly brace